requirements.txt
Django==1.9.12
djangorestframework==3.8.2
